+++ extensions/search-npm/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Search npm Changelog
 
+## [Added a shortcut] - 2025-01-22
+
+- Added a shortcut to copy the package version
+
 ## [Chore & Fixes] - 2025-01-13
 
 - Hide toast when `historyCount` is zero